prodsys/requirements.txt
Django>=5.2
djangorestframework
django-filter
python-dotenv
mysqlclient
